[REF] stock_picking_line_sequence: one file per model
