+++ stock_picking_line_sequence/models/__init__.py
@@ -1,3 +1,5 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import stock
+from . import stock_move
+from . import stock_move_line
+from . import stock_picking
