+++ stock_picking_line_sequence/models/stock_move.py
@@ -1,0 +1,30 @@
+# Copyright 2017 Camptocamp SA - Damien Crier, Alexandre Fayolle
+# Copyright 2017 Eficent Business and IT Consulting Services S.L.
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+from odoo import api, fields, models
+
+
+class StockMove(models.Model):
+    _inherit = "stock.move"
+
+    # re-defines the field to change the default
+    sequence = fields.Integer("HiddenSequence", default=9999)
+
+    # displays sequence on the stock moves
+    sequence2 = fields.Integer(
+        "Sequence",
+        help="Shows the sequence in the Stock Move.",
+        related="sequence",
+        readonly=True,
+        store=True,
+    )
+
+    @api.model
+    def create(self, values):
+        move = super().create(values)
+        # We do not reset the sequence if we are copying a complete picking
+        # or creating a backorder
+        if not self.env.context.get("keep_line_sequence", False):
+            move.picking_id._reset_sequence()
+        return move
