+++ stock_picking_line_sequence/views/stock_picking_views.xml
